tools/requirements.txt
click==8.1.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
